split main, money, and amoeba
